refactor: project structure update #3
