+++ src/component/Details/Cards/NFTCard.jsx
