+++ src/component/Details/Properties/AccordionLayout.jsx
